% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_add_opendata_cols.R
\name{adcp_add_opendata_cols}
\alias{adcp_add_opendata_cols}
\title{Add deployment_id, waterbody, and station columns}
\usage{
adcp_add_opendata_cols(
  dat,
  metadata = NULL,
  deployment_id = NULL,
  waterbody = NULL,
  station = NULL
)
}
\arguments{
\item{dat}{Data frame of ACDP data in long format, as returned by
\code{adcp_pivot_longer()}.}

\item{metadata}{Data frame with metadata information for the deployment in
\code{dat} (e.g., a row from the NSDFA tracking sheet). Must include
columns \code{Depl_ID}, \code{Waterbody}, and \code{Station}. Option to use
default value \code{metadata = NULL} and provide the required values in the
remaining arguments.}

\item{deployment_id}{Unique ID assigned to each deployment. Not used if
\code{metadata} argument is specified.}

\item{waterbody}{Waterbody in which ADCP was deployed. Not used if
\code{metadata} argument is specified.}

\item{station}{Specific area in which ADCP was deployed. Not used if
\code{metadata} argument is specified.}
}
\value{
Returns \code{dat} with columns deployment_id, waterbody, and
  station.
}
\description{
Add deployment_id, waterbody, and station columns
}
