% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_pretreatment_results}
\alias{aggregate_pretreatment_results}
\title{Aggregate results for multiple pretreatments}
\usage{
aggregate_pretreatment_results(
  training.results.i,
  methods.list,
  i,
  pretreatment,
  counter,
  existing.results = NULL
)
}
\arguments{
\item{training.results.i}{Current training results}

\item{methods.list}{List of method names}

\item{i}{Current pretreatment index}

\item{pretreatment}{Vector of pretreatment indices}

\item{counter}{Current counter value}

\item{existing.results}{Existing aggregated results (if any)}
}
\value{
List of aggregated results
}
\description{
Internal function to combine results from multiple pretreatments
}
\keyword{internal}
