% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_cv_control}
\alias{create_cv_control}
\title{Create cross-validation training control}
\usage{
create_cv_control(k.folds = 5, cv.seeds)
}
\arguments{
\item{k.folds}{Number of cross-validation folds}

\item{cv.seeds}{Seeds for cross-validation}
}
\value{
caret trainControl object
}
\description{
Internal function to create consistent caret trainControl objects
}
\keyword{internal}
