% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectra.R
\name{FilterSpectra}
\alias{FilterSpectra}
\title{Filter spectral data frame based on Mahalanobis distance}
\usage{
FilterSpectra(df, filter, return.distances, num.col.before.spectra, window.size)
}
\arguments{
\item{df}{obj a \code{data.frame} object containing columns of spectra and rows of observations. May also contain columns of metadata to the left of the spectra.}

\item{filter}{boolean that determines whether or not the input \code{data.frame} will be filtered.
If \code{TRUE}, \code{df} will be filtered according to squared Mahalanobis distance with a 95% cutoff from a chi-square
distribution with degrees of freedom = number of spectral columns. If \code{FALSE}, a column of squared Mahalanobis
distances \code{h.distance} will be added to the right side of df and all rows will be returned. Default is `TRUE`.}

\item{return.distances}{boolean that determines whether a column of squared Mahalanobis distances will be included
in output \code{data.frame}. If \code{TRUE}, a column of Mahalanobis distances for each row will be added to the right
side of \code{df}. Default is \code{FALSE}.}

\item{num.col.before.spectra}{number of columns to the left of the spectral matrix in \code{df}. Default is 4.}

\item{window.size}{number defining the size of window to use when calculating the covariance of the
spectra (required to calculate Mahalanobis distance). Default is 10.}
}
\value{
If 'filter' is TRUE, returns filtered data frame 'df' and reports the number of rows removed.
      The Mahalanobis distance with a cutoff of 95% of chi-square distribution
      (degrees of freedom = number of wavelengths)is used as filtering criteria.
      If 'filter' is FALSE, returns full input df with column 'h.distances' containing the Mahalanobis
      distance for each row.
}
\description{
Determine Mahalanobis distances of observations (rows) within a given `data.frame`
with spectral data. Option to filter out observations based on these distances.
}
\details{
This function uses a chi-square distribution with 95% cutoff where degrees of
freedom = number of wavelengths (columns) in the input `data.frame`.
}
\examples{
\dontrun{
ikeogu.2017 \%>\%
  dplyr::select(-TCC) \%>\%
  na.omit() \%>\%
  FilterSpectra(df = .,
                filter = T,
                return.distances = F,
                num.col.before.spectra = 5)
}
}
\references{
Johnson, R.A., and D.W. Wichern. 2007. Applied Multivariate Statistical Analysis (6th Edition). pg 189
}
\author{
Jenna Hershberger \email{jmh579@cornell.edu}
}
