% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatCV.R
\name{FormatCV}
\alias{FormatCV}
\title{Format multiple trials with or without overlapping genotypes into training and test sets
according to user-provided cross validation scheme}
\usage{
FormatCV(trial1, trial2, trial3, cv.scheme, seed = NULL,
  remove.genotype = FALSE)
}
\arguments{
\item{trial1}{`data.frame` object that is for use only when `cv.scheme` is provided.
Contains the trial to be tested in subsequent model training functions. The first column
contains unique identifiers, second contains genotypes, third contains reference values,
followed by spectral columns. Include no other columns to right of spectra! Column names
of spectra must start with "X", reference column must be named "reference", and genotype column
must be named "genotype".}

\item{trial2}{`data.frame` object that is for use only when `cv.scheme` is provided.
This data.frame contains a trial that has overlapping genotypes with `trial1`
but that were grown in a different site/year (different environment). Formatting must be
consistent with `trial1`.}

\item{trial3}{`data.frame` object that is for use only when `cv.scheme` is provided.
This data.frame contains a trial that may or may not contain genotypes that overlap with `trial1`.
Formatting must be consistent with `trial1`.}

\item{cv.scheme}{A cross validation (CV) scheme from Jarquín et al., 2017.
Options for cv.scheme include:
*"CV1": untested lines in tested environments
*"CV2": tested lines in tested environments
*"CV0": tested lines in untested environments
*"CV00": untested lines in untested environments}

\item{seed}{Number used in the function `set.seed()` for reproducible randomization.
If `NULL`, no seed is set. Default is `NULL`.}

\item{remove.genotype}{boolean that, if `TRUE`, removes the "genotype" column is removed from
the output `data.frame`. Default is `FALSE`.}
}
\value{
List of data.frames (training set, test set) compiled according to user-provided
cross validation scheme.
}
\description{
Standalone function that is also used within [TrainSpectralModel()]
}
\details{
Use of a cross-validation scheme requires a column in the input `data.frame` named
"genotype" to ensure proper sorting of training and test sets. Variables `trial1` and `trial2`
are required, while `trial 3` is optional.
}
\examples{


}
\references{
Jarquín, D., C. Lemes da Silva, R. C. Gaynor, J. Poland, A. Fritz, R. Howard,
S. Battenfield, and J. Crossa. 2017. Increasing Genomic-Enabled Prediction Accuracy by Modeling
Genotype × Environment Interactions in Kansas Wheat. Plant Genome 10.2
doi:10.3835/plantgenome2016.12.0130
}
