% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mode}
\alias{get_mode}
\title{Internal utility functions}
\usage{
get_mode(vector.input)
}
\arguments{
\item{vector.input}{The mode of this vector of numbers will be calculated
by this function}
}
\value{
mode of the numbers in `vector.input`
}
\description{
Get the mode of a set of numbers. Used in getting summary of
results within [train_spectra()]
}
\keyword{internal}
