\name{haar}
\alias{haar}
\title{
  Haar Wavelet
}
\description{
  The Haar wavelet provides excellent time-domain resolution but poor
  frequency-domain resolution.
}
\usage{haar}
\author{
  Baltasar Trancon y Widemann \email{Baltasar.Trancon@uni-bayreuth.de}
}
\seealso{
  \code{\link{dog}},
  \code{\link{morlet}}
}
\examples{
plot(haar)
}
