% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ikeogu.2017}
\alias{ikeogu.2017}
\title{Example vis-NIRS and reference dataset}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 808 rows and 2158 columns.}
\usage{
ikeogu.2017
}
\description{
The `ikeogu.2017` data set contains raw vis-NIRS scans, total carotenoid content,
and cassava root dry matter content (using the oven method) from the 2017 PLoS One paper by Ikeogu et al.
This dataset contains all original scans but only some of the reference values from the supplementary
files of the paper.
`ikeogu.2017` is a `data.frame` that contains the following columns:
  * study.name = Name of the study as described in Ikeogu et al. (2017).
  * location = Location from which the data was collected
  * year = Year in which data was collected
  * prep.method = Method used for sample preparation. Either "intact" for sliced roots or "mashed" for those blended in a food processor before scanning.
  * DMC.oven = Cassava root dry matter content, the percentage of dry weight relative to fresh weight of a sample after oven drying.
  * TCC = Total carotenoid content (ug/g, unknown whether on a fresh or dry weight basis) as measured by
  high performance liquid chromatography in the studies from CIAT,
  those for which `study.name` starts with "C", and by spectrophotometer in the studies from Umudike, Nigeria,
  those for which `study.name` starts with "U"
  * X350:X2500 = spectral reflectance measured with the QualitySpec Trek: S-10016 vis-NIR spectrometer.
Each cell represents the mean of 150 scans on a single root at a single wavelength.
}
\examples{
\dontrun{
data(ikeogu.2017)

ikeogu.2017[1:10,1:10]

ikeogu.2017 \%>\% group_by(study.name) \%>\% nest() \%>\% mutate(n.obs = map_dbl(data, ~nrow(.)))

ikeogu.2017 \%>\% dplyr::select(-starts_with("X")) \%>\% group_by(study.name) \%>\%
gather(trait, value, c(DMC.oven:TCC), na.rm = T) \%>\%
ggplot(aes(x= study.name, y = value, fill = study.name)) +
facet_wrap(~ trait, scales='free_y', nrow=2) +
geom_boxplot() +
theme_bw()
}
}
\references{
Ikeogu, U.N., F. Davrieux, D. Dufour, H. Ceballos, C.N. Egesi, et al. 2017.
Rapid analyses of dry matter content and carotenoids in fresh cassava roots using a portable
visible and near infrared spectrometer (Vis/NIRS). PLoS One 12(12): 1–17.
doi: 10.1371/journal.pone.0188918.
}
\author{
Original authors:

Original authors: Ikeogu, U.N., F. Davrieux, D. Dufour, H. Ceballos, C.N. Egesi, and J. Jannink.
Curated by Jenna Hershberger
}
\keyword{datasets}
