% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition_data}
\alias{partition_data}
\title{Partition data for training and testing}
\usage{
partition_data(
  df,
  test.data = NULL,
  iteration,
  num.iterations,
  stratified.sampling = TRUE,
  proportion.train = 0.7,
  split.test = FALSE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  train.index = NULL
)
}
\arguments{
\item{df}{Training data frame}

\item{test.data}{Test data frame (optional)}

\item{iteration}{Current iteration number}

\item{num.iterations}{Total number of iterations}

\item{stratified.sampling}{Whether to use stratified sampling}

\item{proportion.train}{Proportion of data to use for training}

\item{split.test}{Whether to split test data}

\item{cv.scheme}{Cross-validation scheme}

\item{trial1, trial2, trial3}{Trial data for CV scheme}

\item{train.index}{Pre-calculated training indices (for stratified sampling)}
}
\value{
List containing train and test data for current iteration
}
\description{
Internal function to handle data partitioning for train_spectra
}
\keyword{internal}
