% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpectra.R
\name{PlotSpectra}
\alias{PlotSpectra}
\title{Plot spectral data, highlighting outliers as identified using Mahalanobis distance}
\usage{
PlotSpectra(input.df, wavelengths, num.col.before.spectra)
}
\arguments{
\item{input.df}{`data.frame` object containing columns of spectra. Spectral columns must be labeled
with an "X" and then the wavelength (example: "X740" = 740nm). Left-most column must be unique ID.
May also contain columns of metadata between the unique ID and spectral columns. Cannot contain
any missing values}

\item{wavelengths}{List of wavelengths (numerical format) represented by each spectral column in `input.df`}

\item{num.col.before.spectra}{Number of columns to the left of the spectral matrix (including unique ID).}
}
\value{
Prints unique ID and metadata for rows identified as outliers. Returns plot of spectral
data with non-outliers in blue and outliers in red. X-axis is wavelengths and y-axis is spectral values.
}
\description{
Generates a ggplot object of given spectra, with wavelength on the x axis and
given spectral values on the y. Mahalanobis distance is used to calculate outliers, which are
both identified on the plot. Rows from the original dataframe are printed to the console for each
outlier that is identified.
}
\examples{
\dontrun{
ikeogu.2017 \%>\%
 rename(unique.id = sample.id) \%>\%
 dplyr::select(unique.id, everything(), -TCC) \%>\%
 na.omit() \%>\%
 PlotSpectra(input.df = .,
             wavelengths = 350:2500,
             num.col.before.spectra = 5)
}
}
\author{
Jenna Hershberger \url{jmh579@cornell.edu}
}
