% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictFromSavedModel.R
\name{PredictFromSavedModel}
\alias{PredictFromSavedModel}
\title{Use provided model object to predict trait values with input dataset}
\usage{
PredictFromSavedModel(input.data, model.stats.location, model.location,
  wavelengths = 740:1070, model.method = "pls")
}
\arguments{
\item{input.data}{`data.frame` object of spectral data for input into a spectral prediction model.
First column contains unique identifiers followed by spectral columns. Include no other columns to
right of spectra! Column names of spectra must start with "X".}

\item{model.stats.location}{String containing file path (including file name) to save location
of "(model.name)_stats.csv" as output from the SaveModel function.}

\item{model.location}{String containing file path (including file name) to location where the
trained model ("(model.name).Rds") was saved as output by the SaveModel function.}

\item{wavelengths}{List of wavelengths represented by each column in `input.data`}

\item{model.method}{Model type to use for training. Options include:
*Partial least squares regression, "pls" (Default)
*Random forest, "rf"
*Support vector machine with linear kernel, "svmLinear"
*Support vector machine with radial kernel, "svmRadial"}
}
\value{
`data.frame` object of predictions for each sample (row). First column is unique identifier
supplied by `input.data` and second is predicted values
}
\description{
#TODO
}
\details{
#TODO
}
