% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_pretreatment_data}
\alias{process_pretreatment_data}
\title{Process pretreatment data for test_spectra}
\usage{
process_pretreatment_data(
  df.list,
  methods.list,
  i,
  n.train,
  n.test,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL
)
}
\arguments{
\item{df.list}{List of pretreated data frames}

\item{methods.list}{List of method names}

\item{i}{Current pretreatment index}

\item{n.train}{Number of training observations}

\item{n.test}{Number of test observations}

\item{cv.scheme}{CV scheme (if any)}

\item{trial1, trial2, trial3}{Trial data frames (if using cv.scheme)}
}
\value{
List containing processed training and test data
}
\description{
Internal function to handle pretreatment data processing and extraction
}
\keyword{internal}
