% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainSpectralModel.R
\name{TrainSpectralModel}
\alias{TrainSpectralModel}
\title{Train a model based predict reference values with spectral data}
\usage{
TrainSpectralModel(
  df,
  num.iterations,
  test.data = NULL,
  k.folds = 5,
  tune.length = 50,
  model.method = "pls",
  best.model.metric = "RMSE",
  stratified.sampling = TRUE,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  split.test = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{\code{data.frame} object. First column contains unique identifiers,
second contains reference values, followed by spectral columns. Include no
other columns to right of spectra! Column names of spectra must start with
"X" and reference column must be named "reference"}

\item{num.iterations}{Number of training iterations to perform}

\item{test.data}{\code{data.frame} with same specifications as \code{df}. Use
if specific test set is desired for hyperparameter tuning. If \code{NULL},
function will automatically train with a stratified sample of 70\%. Default
is \code{NULL}.}

\item{k.folds}{Number indicating the number of folds for k-fold cross-validation
during model training. Default is 5.}

\item{tune.length}{Number delineating search space for tuning of the PLSR
hyperparameter \code{ncomp}. Must be set to 5 when using the random forest
algorithm (\code{model.method == rf}). Default is 50.}

\item{model.method}{Model type to use for training. Valid options include:
\itemize{ \item "pls": Partial least squares regression (Default) \item
"rf": Random forest \item "svmLinear": Support vector machine with linear
kernel \item "svmRadial": Support vector machine with radial kernel }}

\item{best.model.metric}{Metric used to decide which model is best. Must be
either "RMSE" or "Rsquared"}

\item{stratified.sampling}{If \code{TRUE}, training and test sets will be
selected using stratified random sampling. This term is only used if
\code{test.data == NULL}. Default is \code{TRUE}.}

\item{cv.scheme}{A cross validation (CV) scheme from Jarquín et al., 2017.
Options for \code{cv.scheme} include:
\itemize{
    \item "CV1": untested lines in tested environments
    \item "CV2": tested lines in tested environments
    \item "CV0": tested lines in untested environments
    \item "CV00": untested lines in untested environments
}}

\item{trial1}{\code{data.frame} object that is for use only when
\code{cv.scheme} is provided. Contains the trial to be tested in subsequent
model training functions. The first column contains unique identifiers,
second contains genotypes, third contains reference values, followed by
spectral columns. Include no other columns to right of spectra! Column
names of spectra must start with "X", reference column must be named
"reference", and genotype column must be named "genotype".}

\item{trial2}{\code{data.frame} object that is for use only when
\code{cv.scheme} is provided. This data.frame contains a trial that has
overlapping genotypes with \code{trial1} but that were grown in a different
site/year (different environment). Formatting must be consistent with
\code{trial1}.}

\item{trial3}{\code{data.frame} object that is for use only when
\code{cv.scheme} is provided. This data.frame contains a trial that may or
may not contain genotypes that overlap with \code{trial1}. Formatting must
be consistent with \code{trial1}.}

\item{split.test}{boolean that allows for a fixed training set and a split
test set. Example// train model on data from two breeding programs and a
stratified subset (70\%) of a third and test on the remaining samples
(30\%)  of the third. If \code{FALSE}, the entire provided test set
\code{test.data} will remain as a testing set or if none is provided, 30\%
of the provided \code{train.data} will be used for testing. Default is
\code{FALSE}.}

\item{verbose}{If \code{TRUE}, the number of rows removed through filtering
will be printed to the console. Default is \code{TRUE}.}

\item{vector.input}{The mode of this vector of numbers will be calculated by this function}
}
\value{


mode of the numbers in `vector.input`
}
\description{
Trains spectral prediction models using one of several algorithms and sampling
procedures.

Get the mode of a set of numbers. Used in getting summary of results
within [TrainSpectralModel()]
}
\examples{
\donttest{
library(magrittr)
ikeogu.2017 \%>\%
  dplyr::filter(study.name == "C16Mcal") \%>\%
  dplyr::rename(reference = DMC.oven) \%>\%
  dplyr::select(sample.id, reference, dplyr::starts_with("X")) \%>\%
  na.omit() \%>\%
  TrainSpectralModel(df = .,
                     tune.length = 3,
                     num.iterations = 3,
                     best.model.metric = "RMSE",
                     stratified.sampling = TRUE)
}
}
\author{
Jenna Hershberger \email{jmh579@cornell.edu}
}
\keyword{internal}
