% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate spectral function inputs}
\usage{
validate_inputs(
  train.data,
  test.data = NULL,
  cv.scheme = NULL,
  trial1 = NULL,
  trial2 = NULL,
  trial3 = NULL,
  model.method = "pls",
  tune.length = 50,
  best.model.metric = "RMSE",
  proportion.train = 0.7
)
}
\arguments{
\item{train.data, test.data, cv.scheme, trial1, trial2, trial3, model.method, tune.length, best.model.metric, proportion.train}{Input parameters to validate}
}
\value{
NULL (errors are thrown as side effects if validation fails)
}
\description{
Internal function to validate common inputs for spectral functions
}
\keyword{internal}
