% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waves-package.R
\docType{package}
\name{waves-package}
\alias{waves}
\alias{waves-package}
\title{waves: Format and Visualize ADCP Wave Data}
\description{
Facilitates wrangling and analysis of Wave data collected as part of the Centre for Marine Applied Research's Coastal Monitoring Program. Compatible sensors: Workhorse Sentinel Acoustic Doppler Current Profilers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dempsey-CMAR/waves}
  \item Report bugs at \url{https://github.com/dempsey-CMAR/waves/issues}
}

}
\keyword{internal}
