% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_add_opendata_cols.R
\name{wv_add_opendata_cols}
\alias{wv_add_opendata_cols}
\title{Add deployment_id, county, waterbody, and station columns}
\usage{
wv_add_opendata_cols(
  dat,
  metadata = NULL,
  deployment_id = NULL,
  county = NULL,
  waterbody = NULL,
  station = NULL
)
}
\arguments{
\item{dat}{Data frame of wave data.}

\item{metadata}{Data frame with metadata information for the deployment in
\code{dat} (e.g., a row from the ADCP tracking sheet). Must include columns
\code{depl_id}, \code{county}, \code{waterbody}, and \code{station}. Option
to use default value \code{metadata = NULL} and provide the required values
in the remaining arguments.}

\item{deployment_id}{Unique ID assigned to each deployment. Not used if
\code{metadata} argument is specified.}

\item{county}{County in which ADCP was deployed. Not used if \code{metadata}
argument is specified.}

\item{waterbody}{Waterbody in which ADCP was deployed. Not used if
\code{metadata} argument is specified.}

\item{station}{Specific area in which ADCP was deployed. Not used if
\code{metadata} argument is specified.}
}
\value{
Returns \code{dat} with columns deployment_id, county, waterbody, and
  station.
}
\description{
Add deployment_id, county, waterbody, and station columns
}
