% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_variable_names.R
\name{wv_append_long_variable_names}
\alias{wv_append_long_variable_names}
\title{Append "sea_surface_wave_" to height, period, and direction column names}
\usage{
wv_append_long_variable_names(dat)
}
\arguments{
\item{dat}{Data frame as returned from
\code{wv_assign_short_variable_names()}.}
}
\description{
Append "sea_surface_wave_" to height, period, and direction column names
}
