% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_variable_names.R
\name{wv_assign_cf_variable_names}
\alias{wv_assign_cf_variable_names}
\title{Assign climate forecast-like variable names to columns}
\usage{
wv_assign_cf_variable_names(dat)
}
\arguments{
\item{dat}{Data frame as returned from \code{wv_read_txt()}.}
}
\description{
Guidelines for Construction of CF Standard Names:
https://cfconventions.org/Data/cf-standard-names/docs/guidelines.html
}
