% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_variable_names.R
\name{wv_assign_short_variable_names}
\alias{wv_assign_short_variable_names}
\title{Assign short variable names to columns}
\usage{
wv_assign_short_variable_names(dat)
}
\arguments{
\item{dat}{Data frame as returned from \code{wv_assign_cf_variable_names()}.}
}
\description{
Removes "sea_surface_wave_" from wave variable column names.
}
