% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_convert_vars_to_ordered_factor}
\alias{wv_convert_vars_to_ordered_factor}
\title{Convert wave variables to ordered factor to control order of ggplot facets}
\usage{
wv_convert_vars_to_ordered_factor(dat)
}
\arguments{
\item{dat}{Data frame of wave data in long format. Variable names
must be in short format. Flag columns will be dropped.}
}
\value{
Returns \code{dat} with \code{variable} column as an ordered factor.
}
\description{
Convert wave variables to ordered factor to control order of ggplot facets
}
