% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_convert_vars_to_title}
\alias{wv_convert_vars_to_title}
\title{Add a column of wave variables in title case}
\usage{
wv_convert_vars_to_title(dat, convert_to_ordered_factor = TRUE)
}
\arguments{
\item{dat}{Data frame of wave data in long format. Variable names
should be in short format. Flag columns will be dropped.}

\item{convert_to_ordered_factor}{Logical variable indicating whether the new
\code{variable_title} column should be converted to an ordered factor.
Default is \code{TRUE}.}
}
\value{
Returns \code{dat} with an additional \code{variable_title} column
for use in faceted figures.
}
\description{
Add a column of wave variables in title case
}
