% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_create_variable_labels}
\alias{wv_create_variable_labels}
\title{Add a column of axis labels for wave variables}
\usage{
wv_create_variable_labels(dat, convert_to_ordered_factor = TRUE)
}
\arguments{
\item{dat}{Data frame of wave data in long format. Variable names
should be in short format. Flag columns will be dropped.}

\item{convert_to_ordered_factor}{Logical variable indicating whether the new
\code{variable_label} column should be converted to an ordered factor.
Default is \code{TRUE}.}
}
\value{
Returns \code{dat} with an additional \code{variable_label} column
  for use in faceted figures.
}
\description{
Add a column of axis labels for wave variables
}
