% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_extract_deployment_info}
\alias{wv_extract_deployment_info}
\title{Extract deployment date and station name from path of original text file}
\usage{
wv_extract_deployment_info(file_path)
}
\arguments{
\item{file_path}{Path to the file, include file name and extension (.csv or
.txt). File name must include the deployment date and the station name,
separated by " - ", e.g., "2008.09.25 - Coffin Island SW - Waves.txt"}
}
\value{
Returns a tibble with two columns: \code{depl_date}, and
\code{station}.
}
\description{
Extract deployment date and station name from path of original text file
}
