% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_extract_deployment_info2}
\alias{wv_extract_deployment_info2}
\title{Extract deployment date and station name from compiled rds file}
\usage{
wv_extract_deployment_info2(file_path)
}
\arguments{
\item{file_path}{Path to the file, include file name and extension (.rds).
File name must include the deployment date, the station name, and the
deployment id separated by "_", e.g., "2008-09-25_Coffin Island SW_QN004"}
}
\value{
Returns a tibble with three columns: \code{depl_date},
\code{station}, and \code{deployment_id}.
}
\description{
Extract deployment date and station name from compiled rds file
}
