% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_format_report_table.R
\name{wv_format_report_table}
\alias{wv_format_report_table}
\title{Formats tables for typst summary report}
\usage{
wv_format_report_table(report_table, transpose = TRUE)
}
\arguments{
\item{report_table}{Table to include in the summary report.}

\item{transpose}{Logical argument indicating whether to transpose
\code{report_table} before applying the format. Use the default \code{TRUE}
for the deployment table. Set to \code{FALSE} for the document history
table.}
}
\value{
Returns a data frame that will render nicely in the typst report.
  Every cell is converted to text inside square brackets and separated with
  ",", as expected by the typst table function.
}
\description{
Formats tables for typst summary report
}
