% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wv_get_bin_width}
\alias{wv_get_bin_width}
\title{Set histogram bin width for different variables}
\usage{
wv_get_bin_width(var_to_bin)
}
\arguments{
\item{var_to_bin}{Character string of the variable to bin.}
}
\value{
Returns a number that will generate useful histogram for
  \code{var_to_bin}.
}
\description{
Set histogram bin width for different variables
}
