% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_import_data.R
\name{wv_import_data}
\alias{wv_import_data}
\title{Import wave data from rds files.}
\usage{
wv_import_data(path_input = NULL, county = "all")
}
\arguments{
\item{path_input}{Path to the rds files to be imported. Default is the
wave/processed_data/assembled_data folder on the CMAR R drive (user must be
connected to the Perennia VPN).}

\item{county}{Vector of character string(s) indicating the county or counties
for which to import data. The filter is applied to the file path, so the
county name MUST be part of the file path (e.g., the name of the file).
Defaults to all counties.}
}
\description{
Import wave data from rds files.
}
