% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_label_height.R
\name{wv_label_height}
\alias{wv_label_height}
\title{Assign intervals to wave height data}
\usage{
wv_label_height(
  dat,
  column = significant_height_m,
  breaks = NULL,
  label_sep = " "
)
}
\arguments{
\item{dat}{Data frame including the column with observations that will be
assigned to intervals.}

\item{column}{Column in \code{dat} that will be assigned to intervals (NOT
QUOTED).}

\item{breaks}{Vector of breaks.}

\item{label_sep}{Separator for the interval labels ("lower to upper").
Default is " ".}
}
\value{
Returns \code{dat} with an additional column \code{column_labels},
  the interval labels as an ordered factor.
}
\description{
Assigns intervals to \code{column} using the \code{cut()} function.
\code{column} will typically be a \code{wave_height} column. To label
direction data, use \code{wv_label_direction()}.
}
\details{
Intervals start and 0 and end at the maximum value in increments of 0.25 or
0.5
}
