% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_pivot_vars_longer.R
\name{wv_pivot_vars_longer}
\alias{wv_pivot_vars_longer}
\title{Pivot wave data}
\usage{
wv_pivot_vars_longer(dat_wide, vars = NULL)
}
\arguments{
\item{dat_wide}{Data frame of ADCP wave data, as returned from
\code{wv_read_txt()}.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is all variables (with column names in the "short" format).}
}
\value{
Returns data in a long format.
}
\description{
Pivot wave data
}
