% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_flag_vars.R
\name{wv_plot_flags}
\alias{wv_plot_flags}
\title{Plot wave data coloured by flag value}
\usage{
wv_plot_flags(
  dat,
  qc_tests = c("grossrange", "rolling_sd"),
  vars = "all",
  labels = TRUE,
  n_col = NULL,
  flag_title = TRUE,
  plotly_friendly = FALSE
)
}
\arguments{
\item{dat}{Data frame of flagged wave data in long or wide format. Must
include at least one column name with the string "_flag_variable".}

\item{qc_tests}{Character string of QC tests to plot. Default is
\code{qc_tests = c("grossrange", "rolling_sd")}.}

\item{vars}{Character vector of variables to plot. Default is \code{vars =
"all"}, which will make a plot for each recognized variable in \code{dat}.}

\item{labels}{Logical argument indicating whether to convert numeric flag
values to text labels for the legend.}

\item{n_col}{Number of columns for faceted plots.}

\item{flag_title}{Logical argument indicating whether to include a ggtitle of
the qc test and variable plotted.}

\item{plotly_friendly}{Logical argument. If \code{TRUE}, the legend will be
plotted when \code{plotly::ggplotly} is called on \code{p}. Default is
\code{FALSE}, which makes the legend look better in a static figure.}
}
\value{
Returns a list of ggplot objects; one figure for each test in
  \code{qc_tests} and variable in \code{vars}. Points are coloured by the
  flag value and panels are faceted by depth and sensor. faceted by depth and
  sensor.
}
\description{
Plot wave data coloured by flag value
}
