% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_grossrange_flags.R
\name{wv_plot_grossrange_flags}
\alias{wv_plot_grossrange_flags}
\title{Plot sensor_depth_below_surface_m coloured by roc flag.}
\usage{
wv_plot_grossrange_flags(dat, n_col)
}
\arguments{
\item{dat}{Data frame of flagged wave data with variables in long format and
grossrange flags in long or wide format.}

\item{n_col}{Number of columns for the faceted figure.}
}
\value{
ggplot object. Figure shows variables over time, coloured by the
  \code{grossrange_flag_value} column.
}
\description{
Plot sensor_depth_below_surface_m coloured by roc flag.
}
