% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_histogram.R
\name{wv_plot_histogram}
\alias{wv_plot_histogram}
\title{Plot histogram of wave parameters}
\usage{
wv_plot_histogram(
  dat,
  pal = NULL,
  x_axis_label = NULL,
  n_col = 1,
  scales = "fixed",
  binwidth = NULL,
  bar_outline = "grey10",
  show_legend = FALSE
)
}
\arguments{
\item{dat}{Data frame with wave parameter data in wide or long format.}

\item{pal}{Optional colour palette. Default is the "Dark2" palette from
\code{Rcolourbrewer}.}

\item{x_axis_label}{Label for the x-axis. Default is "Wave Height (m)" if
only plotting wave variables and "Wave Period (s)" if only plotting period
variables.}

\item{n_col}{Number of columns for faceted figure. Default is 1.}

\item{scales}{Character string indicating how to treat scales of the faceted
plot. Passed to \code{facet_wrap()}. Default is "fixed". Other options are
"free_x" and "free_y".}

\item{binwidth}{Width of the bins. Passed to \code{geom_histogram()}.}

\item{bar_outline}{Single colour for the outline of the histogram bars.
Default is "grey10". May leave artifacts (lines at the top of some groups)
in the \code{ggplotly()} version of the figure. Use \code{bar_outline = NA}
to avoid.}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{FALSE}.}
}
\value{
Returns a ggplot object.
}
\description{
Plot histogram of wave parameters
}
