% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_wave_params.R
\name{wv_plot_ts}
\alias{wv_plot_ts}
\title{Plot wave parameters over time}
\usage{
wv_plot_ts(
  dat,
  vars = c("Hs", "Hmax"),
  pal = NULL,
  y_axis_label = NULL,
  show_legend = TRUE
)
}
\arguments{
\item{dat}{Data frame with wave parameter data in wide or long format.}

\item{vars}{Character vector indicating which variables in \code{dat} to
plot. Should include only wave height parameters OR only period parameters.}

\item{pal}{Optional colour palette. Default is the "Dark2" palette from
\code{Rcolourbrewer}.}

\item{y_axis_label}{Label for the y-axis. Default is "Wave Height (m)" for
any parameter containing "H" and "Wave Period (s)" for any parameter
containing "T".}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{TRUE}.}
}
\value{
Returns a ggplot object of the plotted variables.
}
\description{
Plot wave parameters over time
}
