% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_summarise_flags.R
\name{wv_summarise_flags}
\alias{wv_summarise_flags}
\title{Generate summary table of flags}
\usage{
wv_summarise_flags(dat, ...)
}
\arguments{
\item{dat}{Data frame of wave data with variables and grossrange ranges in a
long format. Depth flag (\code{depth_trim_flag}) should be in a separate
column.}

\item{...}{Optional argument. Column names (not quoted) from \code{dat} to
use as grouping variables.}
}
\value{
Returns a data frame with the number and percent of observations
assigned each flag value for each variable, qc_test, and groups in
\code{...}.
}
\description{
Generate summary table of flags
}
