% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_test_all.R
\name{wv_test_all}
\alias{wv_test_all}
\title{Apply multiple quality control tests to wave data}
\usage{
wv_test_all(
  dat,
  qc_tests = NULL,
  county = NULL,
  wv_grossrange_table = NULL,
  wv_rolling_sd_table = NULL,
  wv_spike_table = NULL,
  period_hours = 24,
  max_interval_hours = 2,
  align_window = "center",
  keep_sd_cols = FALSE,
  keep_spike_cols = FALSE,
  ping = FALSE
)
}
\arguments{
\item{dat}{Data frame of wave data in a wide format.}

\item{qc_tests}{Character vector of quality control tests to apply to
\code{dat}. Defaults to all available tests: \code{qc_tests =
c("grossrange", "rolling_sd")}.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Used to filter the default \code{wv_grossrange_table}.}

\item{wv_grossrange_table}{Data frame with at least 5 columns:
\code{variable}: entries must match the names of the variables being tested
in \code{dat}; \code{gr_min}: minimum acceptable value; \code{gr_max}:
maximum accepted value ; \code{user_min}: minimum reasonable value;
\code{user_max}: maximum reasonable value.}

\item{wv_rolling_sd_table}{Data frame with at least two columns:
  \code{variable}: must match the names of the variables being tested in
  \code{dat}. \code{rolling_sd_max}: maximum accepted value for the rolling
  standard deviation.

  Default values are used if \code{rolling_sd_table = NULL}. To see the
  default \code{rolling_sd_table}, type \code{subset(wv_thresholds,
  qc_test == "rolling_sd")} in the console.}

\item{wv_spike_table}{Data frame with at least 3 columns: \code{variable}:
  should match the names of the variables being tested in \code{dat}.
  \code{spike_low}: maximum acceptable spike value to "Pass", and
  \code{spike_high}: maximum acceptable value to be flagged "Suspect/Of
  Interest".

  Default values are used if \code{wv_spike_table = NULL}. To see the
  default \code{wv_spike_table}, type \code{subset(wv_thresholds,
  qc_test == "spike")} in the console.}

\item{period_hours}{Length of a full cycle in hours. Default assumes a daily
cycle of \code{period_hours = 24}.}

\item{max_interval_hours}{Maximum accepted interval between two observations.
If the interval between two observations is greater than this value, the
rolling standard deviation will be set to \code{NA}. This is important
because for large intervals, the number of observations in
\code{period_hours} will be small. For example, if samples are collected
every 6 hours, only 4 observations would be used to calculate
\code{roll_sd}.}

\item{align_window}{Alignment for the window used to calculate the rolling
standard deviation. Passed to \code{zoo::rollapply()}. Default is
\code{align_window = "center"}. Other options are \code{"right"} (backward
window) and \code{"left"} (forward window).}

\item{keep_sd_cols}{Logical value. If \code{TRUE}, the columns used to
produce the rolling standard deviation (\code{int_sample}, \code{n_sample},
and \code{sd_roll}) are returned in \code{dat}. Default is \code{FALSE}.}

\item{keep_spike_cols}{Logical value. If \code{TRUE}, the columns used to
produce the spike value are returned in \code{dat}. Default is
\code{FALSE}.}

\item{ping}{Logical argument. If \code{TRUE}, a "ping" sound will be played
when the function has completed. If function is run several times in quick
succession (e.g., for testing the package), this can cause R to abort the
session. Caution is advised when setting this argument to \code{TRUE}.
Default is \code{ping = FALSE}.}
}
\value{
Returns \code{dat} with additional quality control flag columns.
}
\description{
Apply multiple quality control tests to wave data
}
