% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_test_height_period.R
\name{wv_test_crossref}
\alias{wv_test_crossref}
\title{Additional qc for height and period variables}
\usage{
wv_test_crossref(dat)
}
\arguments{
\item{dat}{Data frame of wave variables in wide format.}
}
\value{
Returns \code{dat} with flags for the height and period variables.
}
\description{
Additional qc for height and period variables
}
