% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_test_rate_of_change.R
\name{wv_test_rate_of_change_depth}
\alias{wv_test_rate_of_change_depth}
\title{Flag observations with rapidly changing sensor_depth_below_surface_m}
\usage{
wv_test_rate_of_change_depth(
  dat,
  roc_depth_threshold = 1,
  return_depth_diff = FALSE
)
}
\arguments{
\item{dat}{Dataframe of wave data for a single deployment in wide format.}

\item{roc_depth_threshold}{The change in \code{sensor_depth_below_surface_m}
that will trigger a flag of 3 (in metres). If the first or last row are
assigned a flag of 3, this will be upgraded to a flag of 4. This assumes
that the sensor was recording during deployment or retrieval.}

\item{return_depth_diff}{Logical argument indicating whether to return the
column of \code{depth_diff} = lead(sensor_depth_below_surface_m) -
sensor_depth_below_surface_m.}
}
\value{
Returns \code{dat} with the flag column
  \code{rate_of_change_flag_sensor_depth_below_surface_m} (and optionally
  \code{depth_diff}.
}
\description{
A change in \code{sensor_depth_below_surface_m} greater than
\code{roc_depth_threshold} will trigger a flag of 3 (Suspect/Of Interest).
The flag will be upgraded to 4 (Fail), if it is assigned to the first or last
row. This assumes that the sensor was recording during deployment or
retrieval, and all observations from this timestamp should be filtered out of
the analysis.
}
