% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_test_rolling_sd.R
\name{wv_test_rolling_sd}
\alias{wv_test_rolling_sd}
\title{Apply the rolling standard deviation test to wave variables}
\usage{
wv_test_rolling_sd(
  dat,
  wv_rolling_sd_table = NULL,
  county = NULL,
  vars = NULL,
  period_hours = 24,
  max_interval_hours = 4,
  align_window = "center",
  keep_sd_cols = FALSE
)
}
\arguments{
\item{dat}{Data frame of wave data in wide format.}

\item{wv_rolling_sd_table}{Data frame with at least two columns:
  \code{variable}: must match the names of the variables being tested in
  \code{dat}. \code{rolling_sd_max}: maximum accepted value for the rolling
  standard deviation.

  Default values are used if \code{rolling_sd_table = NULL}. To see the
  default \code{rolling_sd_table}, type \code{subset(wv_thresholds,
  qc_test == "rolling_sd")} in the console.}

\item{county}{Character string indicating the county from which \code{dat}
was collected. Used to filter the default \code{rolling_sd_table}. Not
required if there is a \code{county} column in \code{dat}.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is all variables (with column names in the "short" format).}

\item{period_hours}{Length of a full cycle in hours. Default assumes a daily
cycle of \code{period_hours = 24}.}

\item{max_interval_hours}{Maximum accepted interval between two observations.
If the interval between two observations is greater than this value, the
rolling standard deviation will be set to \code{NA}. This is important
because for large intervals, the number of observations in
\code{period_hours} will be small. For example, if samples are collected
every 6 hours, only 4 observations would be used to calculate
\code{roll_sd}.}

\item{align_window}{Alignment for the window used to calculate the rolling
standard deviation. Passed to \code{zoo::rollapply()}. Default is
\code{align_window = "center"}. Other options are \code{"right"} (backward
window) and \code{"left"} (forward window).}

\item{keep_sd_cols}{Logical value. If \code{TRUE}, the columns used to
produce the rolling standard deviation (\code{int_sample}, \code{n_sample},
and \code{sd_roll}) are returned in \code{dat}. Default is \code{FALSE}.}
}
\value{
Returns \code{dat} in a wide format, with rolling standard deviation
  flag columns for each variable in the form "rolling_sd_flag_variable".
}
\description{
The rolling standard deviation is calculated using the following algorithm:
}
\details{
1. Calculate the interval between the current and previous observation
(\code{int_sample}; in minutes).

2. Determine the number of samples in the specified time frame
(\code{n_sample = round((60 / int_sample) * period_hours)})

3. If \code{int_sample} is really large (> \code{max_interval_hours}), then
set the effective sample number (\code{n_sample_effective}) to 0.

4. Calculate the rolling standard deviation, with the width =
\code{n_sample_effective} and alignment specified by \code{align_window}.
\code{sd_roll} is rounded to 2 decimal places.


When int_sample = period_hours, n_sample = 1, so sd_roll = NA.
}
\seealso{
Other tests: 
\code{\link{wv_test_spike}()}
}
\concept{tests}
