% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_threshold_tables.R
\docType{data}
\name{wv_thresholds}
\alias{wv_thresholds}
\title{Threshold tables}
\format{
Default thresholds for adding flags to wave data.
}
\usage{
wv_thresholds
}
\description{
Threshold tables
}
\keyword{datasets}
