% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_write_report_table.R
\name{wv_write_report_table_nsdfa}
\alias{wv_write_report_table_nsdfa}
\title{Writes deployment table for summary report from NSDFA tracking}
\usage{
wv_write_report_table_nsdfa(metadata)
}
\arguments{
\item{metadata}{Deployment metadata from the NSDFA tracking sheet.}
}
\value{
Returns a tibble with columns for the report table.
}
\description{
Writes deployment table for summary report from NSDFA tracking
}
