% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoPreprocessing.R
\name{DoPreprocessing}
\alias{DoPreprocessing}
\title{Preproccess spectral data according to user-designated method}
\usage{
DoPreprocessing(df, test.data = NULL, preprocessing.method = 1,
  wavelengths = 740:1070)
}
\arguments{
\item{df}{`data.frame` object containing spectral data. First column(s) (optional)
include metadata (with or without reference value column) followed by spectral columns.
Spectral column names must be formatted as "X" followed by wavelength
Include no other columns to right of spectra! No missing values permitted.}

\item{test.data}{`data.frame` object with same format as train.data. Will be appended to df}

\item{preprocessing.method}{Number or list of numbers 1:13 corresponding to desired pretreatment method(s):
*1 = raw data (default)
*2 = standard normal variate (SNV)
*3 = SNV and first derivative
*4 = SNV and second derivative
*5 = first derivative
*6 = second derivative
*7 = Savitzky–Golay filter (SG)
*8 = SNV and SG
*9 = gap segment derivative (window size = 11)
*10 = SG and first derivative (window size = 5)
*11 = SG and first derivative (window size = 11)
*12 = SG and second derivative (window size = 5)
*13 = SG and second derivative (window size = 11)}

\item{wavelengths}{List of wavelengths represented by each column in `df`}
}
\value{
Preprocessed `df` (or list of `data.frames`) with reference column intact
}
\description{
#TODO
}
