% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpectra.R
\name{FilterSpectra}
\alias{FilterSpectra}
\title{Filter spectral data frame based on Mahalanobis distance}
\usage{
FilterSpectra(df, filter, return.distances, num.col.before.spectra, window.size)
}
\arguments{
\item{df}{obj a `data.frame` object containing columns of spectra and rows of observations.
May also contain columns of metadata to the left of the spectra.}

\item{filter}{boolean that determines whether or not the input `data.frame` will be filtered.
If `TRUE`, `df` will be filtered according to Mahalanobis distance with a 95% cutoff from a chi-square
distribution with degrees of freedom = number of spectral columns. If `FALSE`, a column of Mahalanobis
distances `h.distance` will be added to the right side of df and all rows will be returned. Default is `TRUE`.}

\item{return.distances}{boolean that determines whether a column of Mahalanobis distances will be included in output `data.frame`.
If `TRUE`, a column of Mahalanobis distances for each row will be added to the right side of `df`. Default is `FALSE`.}

\item{num.col.before.spectra}{number of columns to the left of the spectral matrix in `df`. Default is 4.}

\item{window.size}{number defining the size of window to use when calculating the covariance of the
spectra (required to calculate Mahalanobis distance). Default is 10.}
}
\value{
If 'filter' is TRUE, returns filtered data frame 'df' and reports the number of rows removed.
      The Mahalanobis distance with a cutoff of 95% of chi-square distribution
      (degrees of freedom = number of wavelengths)is used as filtering criteria.
      If 'filter' is FALSE, returns full input df with column 'h.distances' containing the Mahalanobis
      distance for each row.
}
\description{
Determine Mahalanobis distances of observations (rows) within a given `data.frame`
with spectral data. Option to filter out observations based on these distances.
}
\details{
This function uses a chi-square distribution with 95% cutoff where degrees of
freedom = number of wavelengths (columns) in the input `data.frame`.
}
\references{
Johnson, R.A., and D.W. Wichern. 2007. Applied Multivariate Statistical Analysis (6th Edition). pg 189
}
\author{
Jenna Hershberger \url{jmh579@cornell.edu}
}
