% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveModel.R
\name{SaveModel}
\alias{SaveModel}
\title{Save spectral prediction model and model performance statistics}
\usage{
SaveModel(df, preprocessing = F, save.model = T,
  model.save.folder = NULL, model.name = "PredictionModel",
  best.model.metric = "RMSE", tune.length = 50, model.method = "pls",
  num.iterations = 10, wavelengths = 740:1070,
  stratified.sampling = T, cv.scheme = NULL, trial1 = NULL,
  trial2 = NULL, trial3 = NULL)
}
\arguments{
\item{df}{`data.frame` object. First column contains unique identifiers, second contains
reference values, followed by spectral columns. Include no other columns to right of spectra!
Column names of spectra must start with "X" and reference column must be named "reference"}

\item{preprocessing}{If `TRUE`, 12 preprocessing methods will be applied and their performance
analyzed. If `FALSE`, input data is analyzed as is (raw). Default is `FALSE`.}

\item{save.model}{If `TRUE`, the trained model will be saved in .Rds format to the
location specified by `model.save.folder`. If `FALSE`, model will be output by function
but will not save to file. Default is `TRUE`.}

\item{model.save.folder}{Path to folder where model will be saved. If not provided,
will save to working directory.}

\item{model.name}{Name that model will be saved as in `model.save.folder`. Default is 'PredictionModel'.}

\item{best.model.metric}{Metric used to decide which model is best. Must be either "RMSE" or "Rsquared"}

\item{tune.length}{Number deliniating search space for tuning of the PLSR hyperparameter `ncomp`.
Default is 50.}

\item{model.method}{Model type to use for training. Valid options include:
*"pls": Partial least squares regression (Default)
*"rf": Random forest
*"svmLinear": Support vector machine with linear kernel
*"svmRadial": Support vector machine with radial kernel}

\item{num.iterations}{Number of training iterations to perform}

\item{wavelengths}{List of wavelengths represented by each column in `df`}

\item{stratified.sampling}{If `TRUE`, training and test sets will be selected using stratified
random sampling. This term is only used if `test.data = NULL`. Default is `TRUE`.}

\item{cv.scheme}{A cross validation (CV) scheme from Jarquín et al., 2017.
Options for cv.scheme include:
*"CV1": untested lines in tested environments
*"CV2": tested lines in tested environments
*"CV0": tested lines in untested environments
*"CV00": untested lines in untested environments}

\item{trial1}{`data.frame` object that is for use only when `cv.scheme` is provided.
Contains the trial to be tested in subsequent model training functions. The first column
contains unique identifiers, second contains genotypes, third contains reference values,
followed by spectral columns. Include no other columns to right of spectra! Column names
of spectra must start with "X", reference column must be named "reference", and genotype column
must be named "genotype".}

\item{trial2}{`data.frame` object that is for use only when `cv.scheme` is provided.
This data.frame contains a trial that has overlapping genotypes with `trial1`
but that were grown in a different site/year (different environment). Formatting must be
consistent with `trial1`.}

\item{trial3}{`data.frame` object that is for use only when `cv.scheme` is provided.
This data.frame contains a trial that may or may not contain genotypes that overlap with `trial1`.
Formatting must be consistent with `trial1`.}
}
\value{
List of model stats (in `data.frame`) and trained model object. Saves both to
`model.save.folder` as well. To use optimally trained model for predictions, extract `$finalModel`
and use tuned parameters from `$bestTune`
}
\description{
Saves model and model statistics to model.save.folder as model.name.Rds
and model.name_stats.csv respectively
}
\details{
Wrapper that uses [DoPreprocessing()], [FormatCV()] and [TrainSpectralModel()] functions.
}
\author{
Jenna Hershberger
}
