% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_pivot_flags_longer.R
\name{wv_pivot_flags_longer}
\alias{wv_pivot_flags_longer}
\title{Pivot flagged wave data longer by variable}
\usage{
wv_pivot_flags_longer(dat, qc_tests = NULL, vars = NULL)
}
\arguments{
\item{dat}{Data frame of flagged wave data in wide format.}

\item{qc_tests}{Quality control tests included in \code{dat_wide}. If
\code{dat_wide} only includes the max flag, use \code{qc_tests = "qc"}.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is all variables (with column names in the "short" format). Only
required if variables in \code{dat} are in a wide format.}
}
\value{
Returns \code{dat_wide}, with variables and flags pivoted to a long
format.
}
\description{
Pivot flagged wave data longer by variable
}
