% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_pivot_longer.R
\name{wv_pivot_vars_longer}
\alias{wv_pivot_vars_longer}
\title{Pivot wave data}
\usage{
wv_pivot_vars_longer(dat, first_pivot_col = 3, vars = NULL)
}
\arguments{
\item{dat}{Data frame of ADCP wave data, as returned from
\code{wv_read_txt()}.}

\item{first_pivot_col}{Numeric value indicating the first value to pivot.
Will be ignored if \code{vars} is not \code{NULL}.}

\item{vars}{Vector of character strings indicating which columns to pivot.
Default is first_pivot_col column to the last column.}
}
\value{
Returns data in a long format.
}
\description{
Pivot wave data
}
