% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_height_rose.R
\name{wv_plot_height_rose}
\alias{wv_plot_height_rose}
\title{Generate current rose}
\usage{
wv_plot_height_rose(
  dat,
  breaks = NULL,
  height_cols = NULL,
  height_label = "Wave Height (m)"
)
}
\arguments{
\item{dat}{Data frame with one column name that includes the string
"significant_height_m" and one column name that includes the string
"direction". Other columns will be ignored.}

\item{breaks}{Number of break points for current speed OR a vector of breaks.
Lower-inclusive. Default is 0 to the maximum height by 0.5 m or 0.25 m
increments.}

\item{height_cols}{Vector of colours. Must be the same length as
\code{breaks}. Default is the magma palette from \code{viridis}.}

\item{height_label}{Title of the current speed legend. Default is "Wave
Height (m)".}
}
\value{
Returns an "openair" object, a rose plot of current speed and
  direction.
}
\description{
Generate current rose
}
\details{
Generates a wave height rose using the \code{windRose()} function
  from the \code{openair} package. See help files for
  \code{openair::windRose} for more detail.
}
