% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_wave_params.R
\name{wv_plot_histogram}
\alias{wv_plot_histogram}
\title{Plot histogram of wave parameters}
\usage{
wv_plot_histogram(
  dat,
  vars,
  binwidth = NULL,
  bar_outline = "grey10",
  pal = NULL,
  show_legend = TRUE
)
}
\arguments{
\item{dat}{Data frame with wave parameter data in wide or long format.}

\item{vars}{Character vector indicating which variables in \code{dat} to
plot. Should include only wave height parameters OR only period parameters.}

\item{binwidth}{Width of the bins. Passed to \code{geom_histogram()}.}

\item{bar_outline}{Single colour for the outline of the histogram bars.
Default is "grey10". May leave artifacts (lines at the top of some groups)
in the \code{ggplotly()} version of the figure. Use \code{bar_outline = NA}
to avoid.}

\item{pal}{Optional colour palette. Default is the "Dark2" palette from
\code{Rcolourbrewer}.}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{TRUE}.}
}
\value{
Returns a ggplot object.
}
\description{
Plot histogram of wave parameters
}
