% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_ts.R
\name{wv_plot_ts}
\alias{wv_plot_ts}
\title{Plot wave parameters over time}
\usage{
wv_plot_ts(
  dat,
  pal = NULL,
  y_axis_label = NULL,
  n_col = 1,
  scales = "fixed",
  show_legend = FALSE
)
}
\arguments{
\item{dat}{Data frame with wave parameter data in wide or long format.}

\item{pal}{Optional colour palette. Default is the "Dark2" palette from
\code{Rcolourbrewer}.}

\item{y_axis_label}{Label for the y-axis. Default is "Wave Height (m)" if
only plotting wave variables and "Wave Period (s)" if only plotting period
variables.}

\item{n_col}{Number of columns for faceted figure. Default is 1.}

\item{scales}{Character string indicating how to treat scales of the faceted
plot. Passed to \code{facet_wrap()}. Default is "fixed". Other options are
"free_x" and "free_y".}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{FALSE}.}
}
\value{
Returns a ggplot object of the plotted variables.
}
\description{
Plot wave parameters over time
}
