% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_plot_ts.R
\name{wv_plot_ts}
\alias{wv_plot_ts}
\title{Plot wave parameters over time}
\usage{
wv_plot_ts(
  dat,
  vars = NULL,
  pal = NULL,
  n_col = 1,
  scales = "fixed",
  show_legend = FALSE
)
}
\arguments{
\item{dat}{Data frame with wave parameter data in wide or long format.}

\item{vars}{Character vector of variables to pivot if \code{dat} is in a wide
format. Default is all variables.}

\item{pal}{Optional colour palette. Default is c("#104862", "#62AECA",
"#FFD118") (Indigo dye, Moonstone, Jonquil). For plots with more than 3
variables, these colours are interpolated using
\code{grDevices::colorRampPalette}.}

\item{n_col}{Number of columns for faceted figure. Default is 1.}

\item{scales}{Character string indicating how to treat scales of the faceted
plot. Passed to \code{facet_wrap()}. Default is "fixed". Other options are
"free_x" and "free_y".}

\item{show_legend}{Logical argument indicating whether to show figure legend.
Default is \code{FALSE}.}
}
\value{
Returns a ggplot object of the plotted variables.
}
\description{
Plot wave parameters over time
}
