% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_test_grossrange.R
\name{wv_test_grossrange}
\alias{wv_test_grossrange}
\title{Apply grossrange test to a pair of wave height and period variables}
\usage{
wv_test_grossrange(dat, height_var, period_var)
}
\arguments{
\item{dat}{Data frame of wave variables in wide format.}

\item{height_var}{Wave height variable for which to apply the grossrange test
(in quotes).}

\item{period_var}{Period variable for which to apply the grossrange test (in
quotes).}
}
\value{
Returns \code{dat} with an additional grossrange_flag column for the
  specified wave height and period variables.
}
\description{
The grossrange test is applied to paired wave height and period variables
(e.g., significant_height_m with peak_period_s;
average_height_largest_10_percent_m with period_largest_10_percent_s;
average_height_largest_33_percent_m with period_largest_33_percent_s;
maximum_height_m with period_maximum_s).
}
\details{
Wave period variables with values <= 0 are flagged as "Fail". Based on the
data files, it is likely that values of 0 and -0.1 are used as \code{NULL}
values. All other values are flagged as "Pass".

For the zero-crossing parameters (average_height_largest_33_percent_m,
average_height_largest_10_percent_m, maximum_height_m and corresponding
periods), when wave period is <= 0, the corresponding wave height is almost
always 0 m. The converse is also true (when wave height is 0, the period is
<= 0). This suggests that 0 is used as a NULL value for wave height variables
in the WavesMon software.

For the variables calculated from frequency (significant_height_m,
peak_period_s), wave height of 0 m more often corresponds to a non-zero
frequency.

Wave height parameters are flagged as "Fail" if the wave height is 0 m *and*
the period is <= 0 s. All other variables are flagged as "Pass".
}
