% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wv_write_report_table.R
\name{wv_write_report_table}
\alias{wv_write_report_table}
\title{Writes deployment table for summary report from ADCP TRACKING}
\usage{
wv_write_report_table(metadata)
}
\arguments{
\item{metadata}{Deployment metadata from the ADCP TRACKING "Deployment
Details" sheet.}
}
\value{
Returns a tibble with columns for the report table.
}
\description{
Writes deployment table for summary report from ADCP TRACKING
}
